Start Signal Events tests refactoring
