+++ kie-wb-common-stunner/kie-wb-common-stunner-sets/kie-wb-common-stunner-bpmn/kie-wb-common-stunner-bpmn-backend/src/test/java/org/kie/workbench/common/stunner/bpmn/backend/service/diagram/Marshalling/BPMNDiagramMarshallerBase.java
@@ -127,7 +127,7 @@ public abstract class BPMNDiagramMarshallerBase {
     protected BPMNDiagramMarshaller oldMarshaller;
     protected BPMNDirectDiagramMarshaller newMarshaller;
 
-    public BPMNDiagramMarshallerBase() {
+    protected void init() {
         initMocks(this);
         definitionManager = mock(DefinitionManager.class);
         adapterManager = mock(AdapterManager.class);
