+++ kie-wb-common-stunner/kie-wb-common-stunner-sets/kie-wb-common-stunner-bpmn/kie-wb-common-stunner-bpmn-backend/src/test/java/org/kie/workbench/common/stunner/bpmn/backend/service/diagram/Marshalling/events/StartEvent.java
@@ -16,11 +16,18 @@
 
 package org.kie.workbench.common.stunner.bpmn.backend.service.diagram.Marshalling.events;
 
+import java.util.Arrays;
+import java.util.List;
+
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
 import org.kie.workbench.common.stunner.bpmn.backend.service.diagram.Marshalling.BPMNDiagramMarshallerBase;
+import org.kie.workbench.common.stunner.bpmn.backend.service.diagram.Marshalling.Marshaller;
 import org.kie.workbench.common.stunner.bpmn.definition.BaseStartEvent;
 import org.kie.workbench.common.stunner.bpmn.definition.property.dataio.AssignmentsInfo;
 import org.kie.workbench.common.stunner.bpmn.definition.property.dataio.DataIOSet;
 import org.kie.workbench.common.stunner.bpmn.definition.property.general.BPMNGeneralSet;
+import org.kie.workbench.common.stunner.core.definition.service.DiagramMarshaller;
 import org.kie.workbench.common.stunner.core.diagram.Diagram;
 import org.kie.workbench.common.stunner.core.diagram.Metadata;
 import org.kie.workbench.common.stunner.core.graph.Graph;
@@ -29,12 +36,48 @@ import org.kie.workbench.common.stunner.core.graph.content.definition.Definition
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
+import static org.kie.workbench.common.stunner.bpmn.backend.service.diagram.Marshalling.Marshaller.NEW;
+import static org.kie.workbench.common.stunner.bpmn.backend.service.diagram.Marshalling.Marshaller.OLD;
 
+@RunWith(Parameterized.class)
 public abstract class StartEvent extends BPMNDiagramMarshallerBase {
 
     protected static final String EMPTY_VALUE = "";
     protected static final boolean NON_INTERRUPTING = false;
     protected static final boolean INTERRUPTING = true;
+
+    protected DiagramMarshaller<Graph, Metadata, Diagram<Graph, Metadata>> marshaller = null;
+
+    @Parameterized.Parameters
+    public static List<Object[]> marshallers() {
+        return Arrays.asList(new Object[][] {
+                {OLD}, {NEW}
+        });
+    }
+
+    public StartEvent(Marshaller marshallerType) {
+        super.init();
+        switch(marshallerType) {
+            case OLD: marshaller = oldMarshaller; break;
+            case NEW: marshaller = newMarshaller; break;
+        }
+    }
+
+    abstract void testUnmarshallTopLevelEventFilledProperties() throws Exception;
+
+    abstract void testUnmarshallTopLevelEmptyEventProperties() throws Exception;
+
+    abstract void testUnmarshallSubprocessLevelEventFilledProperties() throws Exception;
+
+    abstract void testUnmarshallSubprocessLevelEventEmptyProperties() throws Exception;
+
+    abstract void testMarshallTopLevelEventFilledProperties() throws Exception;
+
+    abstract void testMarshallTopLevelEmptyEventProperties() throws Exception;
+
+    abstract void testMarshallSubprocessLevelEventFilledProperties() throws Exception;
+
+    abstract void testMarshallSubprocessLevelEventEmptyProperties() throws Exception;
 
     protected <T extends BaseStartEvent> void assertNodesEqualsAfterMarshalling(Diagram<Graph, Metadata> before, Diagram<Graph, Metadata> after, String nodeId, Class<T> startType) {
         T nodeBeforeMarshalling = getStartNodeById(before, nodeId, startType);
